Refactored comparision service
