+++ models/models.go
@@ -1,12 +1,5 @@
 package models
 
-import (
-	"encoding/json"
-	"fmt"
-
-	"github.com/oreuta/easytrip/clients"
-)
-
 //MainRequest gives fields for request from main page
 type MainRequest struct {
 	Currency []string
@@ -24,17 +17,3 @@ type CurrencyBank struct {
 
 //CurrencyBanks is an array of CurrencyBank
 type CurrencyBanks []CurrencyBank
-
-func GetBanks() (bank CurrencyBanks, err error) {
-	client := clients.New()
-	body, err := clients.BankUAClient.Get(client)
-	if err != nil {
-		return nil, fmt.Errorf("ERROR: %v", err)
-	}
-
-	err = json.Unmarshal(body, &bank)
-	if err != nil {
-		fmt.Printf("ERROR: %v", err)
-	}
-	return
-}
